split project.py into MLProjectProcessor and ProjectProcessor
This commit simply moves the two into different files, to have a less
big file, but also to avoid the dynamic imports. These will be removed
in the next commit. This commit is only splitting the files up and
changing the import statements in the notebooks.
